+++ src/wsjcpp_core.h
@@ -130,7 +130,7 @@ public:
   static std::string& trim(std::string& str, const std::string& chars = "\t\n\v\f\r ");
   static std::string toLower(const std::string &str);
   static std::string toUpper(const std::string& str);
-  static void replaceAll(std::string& str, const std::string& from, const std::string& to);
+
 
   static void initRandom();
   static std::string uint2hexString(unsigned int n);
@@ -267,6 +267,7 @@ bool dir_exists(const std::string &dir_path);
 bool file_exists(const std::string &file_path);
 std::vector<std::string> split(const std::string &source, const std::string &delimiter);
 std::string join(const std::vector<std::string> &source, const std::string &delimiter);
+void replace_all_in(std::string &target, const std::string &from, const std::string &to);
 
 } // namespace wsjcpp