+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed #44: Renamed `WsjcppCore::fileExists` to `wsjcpp::file_exists`
 - Fixed #46: Renamed `WsjcppCore::split` to `wsjcpp::split`
 - Renamed `WsjcppCore::join` to `wsjcpp::join`
+- Fixed #47: Rename `WsjcppCore::replaceAll` to `wsjcpp::replace_all`
 
 ## [v0.2.6] - 2026-07-05 (2026 July 5)